requirements.txt
python-telegram-bot[youtube-dlp]==20.7
